moving movie results components into sub folder
